app/requirements.txt
pandas
jupyter
mysql-connector